fix: update loading indicator
