+++ web/screens/Settings/Models/index.tsx
@@ -1,62 +1,28 @@
-import { ChangeEvent, useCallback, useMemo } from 'react'
+import { useCallback, useMemo } from 'react'
 
 import { useDropzone } from 'react-dropzone'
 
-import { Button, Input, ScrollArea } from '@janhq/uikit'
+import { Button, ScrollArea } from '@janhq/uikit'
 
 import { useAtomValue, useSetAtom } from 'jotai'
-import { Plus, SearchIcon, UploadCloudIcon } from 'lucide-react'
+import { Plus, UploadCloudIcon } from 'lucide-react'
 
 import { twMerge } from 'tailwind-merge'
 
-import { useDebouncedCallback } from 'use-debounce'
-
 import useDropModelBinaries from '@/hooks/useDropModelBinaries'
-import { useGetHFRepoData } from '@/hooks/useGetHFRepoData'
 import { setImportModelStageAtom } from '@/hooks/useImportModel'
 
+import ModelSearch from './ModelSearch'
 import RowModel from './Row'
 
-import {
-  importHuggingFaceModelStageAtom,
-  importingHuggingFaceRepoDataAtom,
-} from '@/helpers/atoms/HuggingFace.atom'
 import { downloadedModelsAtom } from '@/helpers/atoms/Model.atom'
+
 const Column = ['Name', 'Model ID', 'Size', 'Version', 'Status', '']
 
 const Models: React.FC = () => {
   const downloadedModels = useAtomValue(downloadedModelsAtom)
   const setImportModelStage = useSetAtom(setImportModelStageAtom)
   const { onDropModels } = useDropModelBinaries()
-  const { loading, getHfRepoData } = useGetHFRepoData()
-
-  const setImportingHuggingFaceRepoData = useSetAtom(
-    importingHuggingFaceRepoDataAtom
-  )
-  const setImportHuggingFaceModelStage = useSetAtom(
-    importHuggingFaceModelStageAtom
-  )
-
-  const debounced = useDebouncedCallback(async (value: string) => {
-    if (value.trim().length === 0) return
-    try {
-      const data = await getHfRepoData(value)
-      setImportingHuggingFaceRepoData(data)
-      setImportHuggingFaceModelStage('REPO_DETAIL')
-    } catch (err) {
-      // TODO: might need to display popup
-      console.error(err)
-    }
-  }, 300)
-
-  const onSearchChanged = useCallback(
-    (e: ChangeEvent<HTMLInputElement>) => {
-      e.preventDefault()
-      e.stopPropagation()
-      debounced(e.target.value)
-    },
-    [debounced]
-  )
 
   const filteredDownloadedModels = useMemo(
     () => downloadedModels.sort((a, b) => a.name.localeCompare(b.name)),
@@ -96,42 +62,7 @@ const Models: React.FC = () => {
       )}
       <div className="m-4 rounded-xl border border-border shadow-sm">
         <div className="flex flex-row justify-between px-6 py-5">
-          <div className="relative w-1/3">
-            <SearchIcon
-              size={20}
-              className="absolute left-2 top-1/2 -translate-y-1/2 text-muted-foreground"
-            />
-            <div className="flex flex-row items-center space-x-4">
-              <Input
-                placeholder="Search or paste Hugging Face URL"
-                className="pl-8"
-                onChange={onSearchChanged}
-              />
-              {loading && (
-                <svg
-                  aria-hidden="true"
-                  role="status"
-                  className="btn-loading-circle"
-                  xmlns="http://www.w3.org/2000/svg"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                >
-                  <circle
-                    cx="12"
-                    cy="12"
-                    r="10"
-                    stroke="currentColor"
-                    strokeWidth="4"
-                  ></circle>
-                  <path
-                    fill="currentColor"
-                    d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"
-                  ></path>
-                </svg>
-              )}
-            </div>
-          </div>
-
+          <ModelSearch />
           <Button
             themes={'outline'}
             className="space-x-2"
